2lab/project/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
python-jose[cryptography]
passlib[bcrypt]
python-dotenv
celery[redis]
redis
pydantic-settings